refactor(view-engine): move runtimeLimits to lib, fix contracts inversion
