+++ packages/view-engine/src/contracts/viewModel.ts
@@ -37,7 +37,7 @@ import type { FilterValidationError } from '../filter/filterModel.js';
 import type {
   RuntimeLimits,
   RuntimeDiagnostic,
-} from '../engine/runtimeLimits.js';
+} from '../lib/runtimeLimits.js';
 import type { ViewHost } from './ViewHost.js';
 
 export type RecordData = Record<string, unknown>;
